backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# PDF processing
pdfplumber==0.11.4
pdf2image==1.17.0
Pillow==10.4.0

# AI inference (llama.cpp Python bindings)
llama-cpp-python==0.3.16

# Data processing & math
pandas==2.2.2
pydantic==2.8.2
